+++ lucene/core/src/java24/org/apache/lucene/internal/vectorization/PanamaVectorizationProvider.java
@@ -78,4 +78,9 @@ final class PanamaVectorizationProvider extends VectorizationProvider {
     }
     return new PostingDecodingUtil(input);
   }
+
+  @Override
+  public BitSetUtil newBitSetUtil() {
+    return PanamaBitSetUtil.INSTANCE;
+  }
 }
